+++ src/RD.Rsse.Engine.VectorSearch/Dto/InternalDocumentIdsWithToken.cs
@@ -1,0 +1,13 @@
+namespace RD.RsseEngine.Dto;
+
+public readonly struct InternalDocumentIdsWithToken(InternalDocumentIdList documentIds, Token token)
+{
+    public readonly InternalDocumentIdList DocumentIds = documentIds;
+
+    public readonly Token Token = token;
+
+    public override string ToString()
+    {
+        return $"InternalDocumentIdsWithToken Token {Token}";
+    }
+}